Fixed bug when starting discovery after reenbled bluetooth
When bluetooth goes off during connecting we go to Idle and request
to turn on bluetooth. If user does so we go back to Connecting and
restart discovery and listening. However, that restart often fails
if we do it immediately after BluetoothOn event because bluetooth seems
to need time to become fully activated. Doing retries on failed
StartDiscovery and StartListening solves the problem.